Move ay_tests to aytests
