+++ bin/aytests
@@ -1,8 +1,6 @@
 #!/usr/bin/env ruby
 
 $LOAD_PATH.unshift File.join(File.dirname(__FILE__), "..", "lib")
-require "bundler/setup"
-Bundler.require(:default)
 require "ay_tests/cli"
 
 ENV["VAGRANT_LOG"] ||= "warn"         # Set Vagrant log level to 'warn'
