+++ kiwi/definitions/autoyast/install_definition.rb
@@ -37,7 +37,7 @@ Veewee::Definition.declare({
     # autoyast profile to the installer. veewee's built in webserver solution
     # doesn't work reliably with autoyast due to some timing issues.
     :before_create => Proc.new do
-      require "../lib/ay_tests/web_server"
+      require "../lib/aytests/web_server"
       require "pathname"
       Thread.new do
         AYTests::WebServer.new(base_dir: Pathname.pwd.join(".."), name: definition.box.name).start
