+++ lib/aytests/image_builder.rb
