+++ lib/aytests/installer.rb
