+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require "pathname"
-require "ay_tests"
+require "aytests"
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
